refactor: new cmd package inside internal
