+++ src/internal/cmd/bootstrap/bootstrap_test.go
