+++ src/internal/cmd/bootstrap/secrets.go
