+++ src/internal/cmd/bootstrap/secrets_test.go
