+++ src/internal/cmd/generate/generator.go
